Seed data and update data type

+++ InsuranceClaimSystem.API/Features/Claims/CreateClaim/CreateClaimRequest.cs
@@ -5,7 +5,7 @@ namespace InsuranceClaimSystem.API.Features.Claims.CreateClaim
 {
     public class CreateClaimRequest : IRequest<Result<Guid>>
     {
-        public CreateClaimRequest(string name, string description, float amount, DateTime createdDate)
+        public CreateClaimRequest(string name, string description, double amount, DateTime createdDate)
         {
             Name = name;
             Description = description;
@@ -17,7 +17,7 @@ namespace InsuranceClaimSystem.API.Features.Claims.CreateClaim
 
         public string Name { get; init; }
         public string Description { get; init; }
-        public float Amount { get; init; }
+        public double Amount { get; init; }
         public DateTime CreatedDate { get; init; }
     }
 }
